Todo con usuario

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
 import { NextResponse } from "next/server";
 import * as yup from "yup";
@@ -41,9 +42,15 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+  const user = await getUserSessionServer();
+
+  if (!user) {
+    return NextResponse.json('NO autorizado', { status: 401 })
+  }
+
   try {
     const { complete, description } = await postSchema.validate(await request.json())
-    const todo = await prisma.todo.create({ data: { complete, description } })
+    const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } })
 
     return NextResponse.json(todo)
 
@@ -54,6 +61,12 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
-  const todo = await prisma.todo.deleteMany({ where: { complete: true } })
+  const user = await getUserSessionServer();
+
+  if (!user) {
+    return NextResponse.json('NO autorizado', { status: 401 })
+  }
+
+  const todo = await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } })
   return NextResponse.json({ message: 'Todo eliminado con exito', todo })
 }
